Reorganização do projeto, separação da API e formulário, melhorias nas práticas de desenvolvimento
