+++ certificados/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+certificados_bp = Blueprint('certificados', __name__)
+
+from . import views
