+++ config.py
@@ -1,0 +1,7 @@
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+class Config:
+    MONGO_URI = os.getenv('MONGO_URI')
